third_party/blink/web_tests/platform/linux/external/wpt/mathml/relations/css-styling/display-2-expected.txt

This is a testharness.js-based test.
PASS block display
PASS block display with contrained width
PASS list display inside display block
PASS inline display
PASS inline-block display
FAIL table display (math) assert_approx_equals: block size expected 80 +/- 1 but got 78
PASS table display (mrow)
FAIL inline-table display (math) assert_approx_equals: block size expected 60 +/- 1 but got 58
PASS inline-table display (mrow)
PASS flexbox display (math)
PASS flexbox display (mrow)
PASS grid display (math)
PASS grid display (mrow)
FAIL ruby display (math) assert_approx_equals: block size expected 80 +/- 1 but got 30
FAIL ruby display (mrow) assert_approx_equals: block size expected 80 +/- 1 but got 45
FAIL block display with column width (math) assert_approx_equals: block size expected 80 +/- 1 but got 45
FAIL block display with column width (mrow) assert_approx_equals: block size expected 80 +/- 1 but got 45
Harness: the test ran to completion.
